requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.0
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.1.0
django-extensions==3.2.3
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
stripe==11.2.0
twilio==9.3.6
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0